Sat, Jul 8, 2023, 3:20 PM +05:00

+++ half/halfs.c
@@ -18,8 +18,8 @@ int main(void)
 // TODO: Complete the function
 float half(float bill, float tax, int tip)
 {
-     bill = bill_amount/2;
-     tax = tax_percent/2;
+     bill = float(bill_amount/2);
+     tax = float(tax_percent/2);
      tip = float(tip_percent/2);
     tax *= bill
     tip *= bill
